python-agent/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.31.0
sniffio==1.3.1
starlette==0.38.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
